Separate components into multiple files
