+++ src/components/PageHome/PageHome.css
@@ -2,8 +2,6 @@
 .nav-container {
     background-color: var(--blue);
     padding: 20px;
-    /* margin: 20px; */
-    /* margin: 0 auto; */
     justify-content: space-around;
 }
 
@@ -18,12 +16,12 @@
 }
 
 .search-form {
-    /* background-color: orange; */
     margin: 0 50px;
     flex-grow: 1;
     max-width: 80%;
     border-radius: 15px;
     overflow: hidden;
+    border: 2px solid black;
 }
 
 .search-bar-container {
@@ -32,17 +30,26 @@
 }
 
 .search-bar {
-  width: 100%;
-  height: 100%;
+    width: 100%;
+    height: 100%;
+    border-top-style: none;
+    border-bottom-style: none;
+    border-right-style: none;
+    border-left: 2px solid black;
+    outline: none;
 }
 
 .search-input {
+    font-size: 1.2rem;
     border: none;
     background-color: var(--light-gray);
     border-radius: 0;
+    padding: 0;
+    margin: 0;
 }
 
-.search-button {
+.search-icon {
+    background-color: white;
     font-size: 1.5rem;
     padding: 10px;
 }
@@ -50,12 +57,9 @@
 .product-grid {
     max-width: 750px;
     margin: 10px auto;
-    /* display: flex;
-    flex-wrap: wrap; */
     display: grid;
     gap: 10px;
     grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-    /* grid-template-columns: repeat(4, 1fr); */
 }
 
 .product-item {
@@ -73,14 +77,8 @@
     max-width: 300px;
 }
 
-/* .product-info {
-    margin: 10px;
-} */
-
-
 .product-name {
     font-size: 1.3rem;
-    /* font-weight: bold; */
     margin: 10px;
 }
 
@@ -127,10 +125,6 @@
     border: 2px solid black;
 }
 
-/* .dropdown-container:hover .dropdown-menu {
-    transform: scaleY(1);
-} */
-
 .dropdown-title {
     color: white;
     font-size: 1.2rem;
@@ -138,15 +132,16 @@
 }
 
 .dropdown-menu {
+    padding: 20px;
     background-color: var(--light-gray);
     border: 2px solid var(--dark-gray);
-    /* border-radius: 0 0 15px 15px; */
     transform-origin: top;
     position: absolute;
     transition: transform 0.2s;
 }
 
 .dropdown-item {
+    font-size: 1.3rem;
     background-color: transparent;
     border: none;
     border-radius: 0;
